+++ apps/linkedin-to-notion/src/contents/shared/waitForElementToExist.ts
